Fix undefined useApi hooks and change defintion return format
