+++ app/src/apiClient/useFlashcardApiClient.js
@@ -1,17 +1,11 @@
 import * as React from "react";
-import { useState, useEffect } from "react";
 
-import useAuth0 from "./useAuth0";
+import useAuth0 from "../auth/useAuth0";
 
 const makeApi = (accessToken) => {
   const actions = {
     getFlashcards: () => _get("/api/flashcards"),
     getCard: (cardId) => _get(`/api/flashcards/${cardId}`),
-    getDefinition: (word) => {
-      console.log("word", word);
-      return _get("/api/dictionary?word=" + word);
-    },
-    addOrUpdateUser: (user) => _post("/api/users", { user }),
     editIsLearnt: (card) => {
       return _put(`/api/flashcards/${card.id}`, { card });
     },
@@ -79,15 +73,15 @@ const makeApi = (accessToken) => {
   return actions;
 };
 
-const userApiClient = () => {
+const useFlashcardApiClient = () => {
   const { isAuthenticated, getAccessTokenSilently } = useAuth0();
-  const [state, setState] = useState({
+  const [state, setState] = React.useState({
     loading: true,
     error: null,
-    apiClient: undefined,
+    flashcardApi: undefined,
   });
 
-  useEffect(() => {
+  React.useEffect(() => {
     if (isAuthenticated) {
       (async () => {
         try {
@@ -95,10 +89,10 @@ const userApiClient = () => {
           setState({
             loading: false,
             error: null,
-            apiClient: makeApi(accessToken),
+            flashcardApi: makeApi(accessToken),
           });
         } catch (error) {
-          setState({ loading: false, error, apiClient: undefined });
+          setState({ loading: false, error, flashcardApi: undefined });
         }
       })();
     }
@@ -107,4 +101,4 @@ const userApiClient = () => {
   return state;
 };
 
-export default userApiClient;
+export default useFlashcardApiClient;
